# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: HelpMate API
  version: 0.1.0
paths:
...
  /api/chat/sessions:
    get:
      tags:
      - chat
      summary: List Sessions
      operationId: list_sessions_api_chat_sessions_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ChatSession'
                type: array
                title: Response List Sessions Api Chat Sessions Get
    post:
      tags:
      - chat
      summary: Create Session
      operationId: create_session_api_chat_sessions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSession'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/chat/sessions/{session_id}/messages:
    get:
      tags:
      - chat
      summary: List Messages
      operationId: list_messages_api_chat_sessions__session_id__messages_get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatMessage'
                title: Response List Messages Api Chat Sessions  Session Id  Messages Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - chat
      summary: Post Message
      operationId: post_message_api_chat_sessions__session_id__messages_post
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostMessageIn'
      responses:
        '200':
          description: 'Server-sent events. Each frame is `event: <type>` + `data: <ChatEvent JSON>`;
            `: ping` comments are heartbeats.'
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/ChatEvent'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/memory/suggestions:
    get:
      tags:
      - memory
      summary: List Suggestions
      operationId: list_suggestions_api_memory_suggestions_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/SuggestionStatus'
          - type: 'null'
          default: pending
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MemorySuggestion'
                title: Response List Suggestions Api Memory Suggestions Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/memory/suggestions/{suggestion_id}/decision:
    post:
      tags:
      - memory
      summary: Decide Suggestion
      operationId: decide_suggestion_api_memory_suggestions__suggestion_id__decision_post
      parameters:
      - name: suggestion_id
        in: path
        required: true
        schema:
          type: string
          title: Suggestion Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SuggestionDecisionIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemorySuggestion'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/memory/facts/{fact_id}:
    delete:
      tags:
      - memory
      summary: Delete Fact
      operationId: delete_fact_api_memory_facts__fact_id__delete
      parameters:
      - name: fact_id
        in: path
        required: true
        schema:
          type: string
          title: Fact Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/auth/login:
    post:
      tags:
      - auth
      summary: Login
      operationId: login_api_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/mfa:
    post:
      tags:
      - auth
      summary: Verify Mfa
      operationId: verify_mfa_api_auth_mfa_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaIn'
        required: true
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ChatMessage:
      properties:
        id:
          type: string
          title: Id
        session_id:
          type: string
          title: Session Id
        role:
          $ref: '#/components/schemas/Role'
        text:
          type: string
          title: Text
        source:
          type: string
          enum:
          - text
          - voice
          title: Source
          default: text
        created_at:
          type: string
          format: date-time
          title: Created At
      additionalProperties: false
      type: object
      required:
      - id
      - session_id
      - role
      - text
      - created_at
      title: ChatMessage
    ChatSession:
      properties:
        id:
          type: string
          title: Id
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        created_at:
          type: string
          format: date-time
          title: Created At
      additionalProperties: false
      type: object
      required:
      - id
      - created_at
      title: ChatSession
...
    CreateSessionIn:
      properties:
        title:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Title
      additionalProperties: false
      type: object
      title: CreateSessionIn
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoginIn:
      properties:
        username:
          type: string
          maxLength: 100
          minLength: 1
          title: Username
        password:
          type: string
          maxLength: 200
          minLength: 1
          title: Password
      additionalProperties: false
      type: object
      required:
      - username
      - password
      title: LoginIn
    LoginOut:
      properties:
        mfa_required:
          type: boolean
          title: Mfa Required
        challenge_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Challenge Id
      additionalProperties: false
      type: object
      required:
      - mfa_required
      title: LoginOut
...
    MemorySuggestion:
      properties:
        id:
          type: string
          title: Id
        text:
          type: string
          title: Text
        source:
          type: string
          title: Source
        status:
          $ref: '#/components/schemas/SuggestionStatus'
          default: pending
        created_at:
          type: string
          format: date-time
          title: Created At
      additionalProperties: false
      type: object
      required:
      - id
      - text
      - source
      - created_at
      title: MemorySuggestion
      description: A fact the agent noticed. Stored as a MemoryFact only if the owner approves it.
...
    MfaIn:
      properties:
        challenge_id:
          type: string
          title: Challenge Id
        code:
          type: string
          pattern: ^\d{6}$
          title: Code
      additionalProperties: false
      type: object
      required:
      - challenge_id
      - code
      title: MfaIn
...
    PostMessageIn:
      properties:
        text:
          type: string
          maxLength: 4000
          minLength: 1
          title: Text
        source:
          type: string
          enum:
          - text
          - voice
          title: Source
          default: text
      additionalProperties: false
      type: object
      required:
      - text
      title: PostMessageIn
    Proposal:
      properties:
        id:
          type: string
          title: Id
        session_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
        tool:
          type: string
          title: Tool
        title:
          type: string
          title: Title
        summary:
          type: string
          title: Summary
        args:
          additionalProperties: true
          type: object
          title: Args
        preview:
          anyOf:
          - type: string
          - type: 'null'
          title: Preview
        risk:
          $ref: '#/components/schemas/Risk'
          default: write
        status:
          $ref: '#/components/schemas/ProposalStatus'
          default: pending
        created_at:
          type: string
          format: date-time
          title: Created At
        decided_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Decided At
        result:
          anyOf:
          - type: string
          - type: 'null'
          title: Result
      additionalProperties: false
      type: object
      required:
      - id
      - tool
      - title
      - summary
      - args
      - created_at
      title: Proposal
      description: A write the agent wants to make. Nothing runs until the owner approves it.
    ProposalStatus:
      type: string
      enum:
      - pending
      - approved
      - rejected
      - executed
      - failed
      title: ProposalStatus
...
    Risk:
      type: string
      enum:
      - write
      - external
      title: Risk
    Role:
      type: string
      enum:
      - user
      - assistant
      title: Role
...
    SuggestionDecisionIn:
      properties:
        decision:
          type: string
          enum:
          - approve
          - reject
          title: Decision
      additionalProperties: false
      type: object
      required:
      - decision
      title: SuggestionDecisionIn
    SuggestionStatus:
      type: string
      enum:
      - pending
      - approved
      - rejected
      title: SuggestionStatus
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    ErrorEvent:
      properties:
        type:
          const: error
          default: error
          title: Type
          type: string
        code:
          title: Code
          type: string
        message:
          title: Message
          type: string
      required:
      - code
      - message
      title: ErrorEvent
      type: object
    MessageDelta:
      properties:
        type:
          const: message.delta
          default: message.delta
          title: Type
          type: string
        text:
          title: Text
          type: string
      required:
      - text
      title: MessageDelta
      type: object
    MessageDone:
      properties:
        type:
          const: message.done
          default: message.done
          title: Type
          type: string
        message_id:
          title: Message Id
          type: string
        ttft_ms:
          anyOf:
          - type: integer
          - type: 'null'
          default: null
          title: Ttft Ms
        tokens:
          anyOf:
          - type: integer
          - type: 'null'
          default: null
          title: Tokens
      required:
      - message_id
      title: MessageDone
      type: object
    ProposalCreated:
      properties:
        type:
          const: proposal.created
          default: proposal.created
          title: Type
          type: string
        proposal:
          $ref: '#/components/schemas/Proposal'
      required:
      - proposal
      title: ProposalCreated
      type: object
    ToolResult:
      properties:
        type:
          const: tool.result
          default: tool.result
          title: Type
          type: string
        call_id:
          title: Call Id
          type: string
        ok:
          title: Ok
          type: boolean
        summary:
          title: Summary
          type: string
      required:
      - call_id
      - ok
      - summary
      title: ToolResult
      type: object
    ToolStarted:
      properties:
        type:
          const: tool.started
          default: tool.started
          title: Type
          type: string
        call_id:
          title: Call Id
          type: string
        tool:
          title: Tool
          type: string
        args:
          additionalProperties: true
          title: Args
          type: object
      required:
      - call_id
      - tool
      - args
      title: ToolStarted
      type: object
    ChatEvent:
      discriminator:
        mapping:
          error: '#/components/schemas/ErrorEvent'
          message.delta: '#/components/schemas/MessageDelta'
          message.done: '#/components/schemas/MessageDone'
          proposal.created: '#/components/schemas/ProposalCreated'
          tool.result: '#/components/schemas/ToolResult'
          tool.started: '#/components/schemas/ToolStarted'
        propertyName: type
      oneOf:
      - $ref: '#/components/schemas/MessageDelta'
      - $ref: '#/components/schemas/ToolStarted'
      - $ref: '#/components/schemas/ToolResult'
      - $ref: '#/components/schemas/ProposalCreated'
      - $ref: '#/components/schemas/MessageDone'
      - $ref: '#/components/schemas/ErrorEvent'